+++ test/support/domain.ex
@@ -15,5 +15,8 @@ defmodule AshStorage.Test.Domain do
     resource AshStorage.Test.IntegerPost
     resource AshStorage.Test.IntegerAttachment
     resource AshStorage.Test.ExtraAttrsPost
+    resource AshStorage.Test.ChecksumVerifyingPost
+    resource AshStorage.Test.NoHeadPost
+    resource AshStorage.Test.MultipartEtagPost
   end
 end